moved raw codes for browser and server
